+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/hybridakscluster/__init__.py
@@ -10,7 +10,6 @@
 
 from .__cmd_group import *
 from ._list import *
-from ._restart_node import *
 from ._show import *
 from ._update import *
 from ._wait import *